minor fixes plus makefile update

+++ Fillit/fillit.h
@@ -6,7 +6,7 @@
 /*   By: jkuusist <jkuusist@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/26 10:32:01 by jkuusist          #+#    #+#             */
-/*   Updated: 2019/12/16 16:50:00 by lharvey          ###   ########.fr       */
+/*   Updated: 2019/12/16 17:11:32 by jkuusist         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,6 +39,6 @@ void				print_map(t_block **bit_field, unsigned int map_size,
 t_block				**copy_tblocks(t_block **src);
 void				free_tblocks(t_block **array);
 unsigned long long	square_root(unsigned long long x);
-unsigned int		check_four(unsigned *map_field, unsigned int map_size);
+unsigned int		check_four(unsigned short *map_field, unsigned int map_size);
 
 #endif
